+++ docs/conf.py
@@ -40,9 +40,15 @@ intersphinx_mapping = {
     "matplotlib": ("https://matplotlib.org/stable/", None),
     "pandas": ("https://pandas.pydata.org/pandas-docs/stable", None),
     "xarray": ("https://docs.xarray.dev/en/stable/", None),
-    "cartopy": ("https://scitools.org.uk/cartopy/docs/latest", None),
+    "cartopy": ("https://cartopy.readthedocs.io/stable/", None),
     "cftime": ("https://unidata.github.io/cftime/", None),
     "metpy": ("https://unidata.github.io/MetPy/latest/", None),
+    "shapely": ("https://shapely.readthedocs.io/en/stable/", None),
+    "geopandas": ("https://geopandas.org/en/stable/", None),
+    "sklearn": ("https://scikit-learn.org/stable/", None),
+    "scipy": ("https://docs.scipy.org/doc/scipy/", None),
+    "pyproj": ("https://pyproj4.github.io/pyproj/stable/", None),
+    "seaborn": ("https://seaborn.pydata.org/", None),
 }
 
 # myst parser (markdown)